+++ esp-metadata-generated/src/_generated_esp32c5.rs
@@ -388,6 +388,12 @@ macro_rules! property {
     ("uart.ram_size", str) => {
         stringify!(128)
     };
+    ("uart.version") => {
+        2
+    };
+    ("uart.version", str) => {
+        stringify!(2)
+    };
     ("uart.peripheral_controls_mem_clk") => {
         true
     };