+++ esp-storage/src/buffer.rs
@@ -8,6 +8,7 @@ pub fn uninit_slice_mut(bytes: &mut [u8]) -> &mut [MaybeUninit<u8>] {
     unsafe { core::slice::from_raw_parts_mut(bytes.as_mut_ptr() as *mut _, bytes.len()) }
 }
 
+#[cfg_attr(not(feature = "bytewise-read"), expect(unused))]
 pub type FlashWordBuffer = FlashBuffer<4>;
 
 pub type FlashSectorBuffer = FlashBuffer<4096>;